+++ check_manifest.py
@@ -35,7 +35,7 @@ except ImportError:
     import configparser as ConfigParser
 
 
-__version__ = '0.22.dev0'
+__version__ = '0.22'
 __author__ = 'Marius Gedminas <marius@gedmin.as>'
 __licence__ = 'MIT'
 __url__ = 'https://github.com/mgedmin/check-manifest'