+++ CHANGES.rst
@@ -11,6 +11,8 @@ Changelog
 * Fix IOError when files listed under version control are missing (`issue #32
   <https://github.com/mgedmin/check-manifest/issues/32>`__).
 
+* Improved wording of the match/do not match messages (`issue #34
+  <https://github.com/mgedmin/check-manifest/issues/34>`__).
 
 0.21 (2014-06-13)
 -----------------